Loop for PixelAlive is complete -- somehow it never goes into setInjectionSchema

+++ HWDescription/RD53.cc
@@ -8,9 +8,6 @@
 */
 
 #include "RD53.h"
-#include "../Utils/ChannelGroupHandler.h"
-
-#include <unordered_map>
 
 namespace Ph2_HwDescription
 {
@@ -804,11 +801,11 @@ namespace Ph2_HwDescription
 																  cal_aux_delay);
   }
   
-  template<int NBITS>
-  std::bitset<NBITS> RD53::SetBits (unsigned int nBit2Set)
+  template<size_t NBITS>
+  std::bitset<NBITS> RD53::SetBits (size_t nBit2Set)
   {
     std::bitset<NBITS> output(0);
-    for (unsigned int i = 0; i < nBit2Set; i++) output[i] = 1;
+    for (size_t i = 0; i < nBit2Set; i++) output[i] = 1;
     return output;
   }
 }
